Feature2 uses a service from Feature1
The requirements have changed, and a data service that is the
responsibility of the feature1 module is now required in feature2. This
commit therefore places a dependency between the two, so that the shared
service can be re-used.

Of course, there is an argument that this service be moved to core!