Introduces a core module, for each feature
In addition to a feature module (ie my-app.featureone), there is now a
core module. This is the "API layer" for each feature module, and will
contain code that could possibly be shared by other features. This will
ideally be kept as thin as possible.

Doing this means that the main feature module will contain only
feature-specific code.

As these new modules are 'core layer', they depend on my-app.core and are
loaded up-front. Whereas, the remaining of the feature could potentially
be loaded lazily. Due to this, the core layer for a feature will now be
responsible for handling routing for within that module. This gives the
core layer an opportunity to perform lazy loading, in order to resolve the
route.

To show this kind of interaction, the home page now uses both of the
shared services (FeatureOneService and FeatureTwoService). But in order to
do so it can no longer remain in the core layer, instead it has been moved
up to the top level my-app module.